Add files via upload
This PowerShell script is helpful for uploading and downloading a blob from the azure container. This downloads a blob and manipulates the file to reduce the size of a file for uploading. It deletes the existing file in a container and then reuploads it to the container. This can be automated using Task Scheduler as well.